# swagger.yaml
openapi: 3.0.0
info:
  title: MindLoo API
  version: 2.0.0
  description: |
    Complete RESTful API for managing patients, therapists, appointments, and therapy-related operations.
    
    ## Features
    - Patient and Therapist Management
    - Appointment Scheduling
    - Therapy Session Tracking
    - Medical Records Management
    - Secure Messaging System
    - Review and Rating System
    - Treatment Planning and Goal Tracking
    
  contact:
    name: Stefan Ramač
    url: https://www.linkedin.com/in/stefanramac
servers:
  - url: http://localhost:3000
    description: Development server
...
components:
  schemas:
...
    Patient:
      type: object
      required:
        - patientId
        - firstName
        - lastName
        - email
      properties:
        patientId:
          type: string
          example: "P001"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        profile:
          type: object
          properties:
            age:
              type: integer
              example: 30
            gender:
              type: string
              example: "male"
        contactInfo:
          type: object
          properties:
            phone:
              type: string
              example: "+1234567890"
            address:
              type: string
              example: "123 Main St"
            place:
              type: string
              example: "New York"
            country:
              type: string
              example: "USA"
    
    # Therapist Schema
    Therapist:
      type: object
      required:
        - therapistId
        - firstName
        - lastName
        - email
      properties:
        therapistId:
          type: string
          example: "T001"
        firstName:
          type: string
          example: "Jane"
        lastName:
          type: string
          example: "Smith"
        email:
          type: string
          format: email
          example: "jane.smith@mindloo.com"
        type:
          type: string
          default: "therapist"
        profile:
          type: object
          properties:
            age:
              type: integer
              example: 35
            gender:
              type: string
              example: "female"
            specialization:
              type: string
              example: "Cognitive Behavioral Therapy"
            experience:
              type: integer
              example: 10
        contactInfo:
          type: object
          properties:
            phone:
              type: string
            address:
              type: string
            place:
              type: string
            country:
              type: string
    
    # Appointment Schema
    Appointment:
      type: object
      required:
        - therapistId
        - patientId
        - date
        - timeSlot
      properties:
        appointmentId:
          type: string
          readOnly: true
        therapistId:
          type: string
          example: "T001"
        patientId:
          type: string
          example: "P001"
        date:
          type: string
          format: date
          example: "2025-10-25"
        timeSlot:
          type: object
          required:
            - start
            - end
          properties:
            start:
              type: string
              example: "10:00"
            end:
              type: string
              example: "11:00"
        status:
          type: string
          enum: [scheduled, completed, cancelled, rescheduled]
          default: scheduled
        subject:
          type: string
          example: "Initial consultation"
        notes:
          type: string
    
    # Availability Schema
...
    Session:
      type: object
      required:
        - appointmentId
        - therapistId
        - patientId
        - sessionDate
        - duration
      properties:
        sessionId:
          type: string
          readOnly: true
        appointmentId:
          type: string
        therapistId:
          type: string
        patientId:
          type: string
        sessionDate:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in minutes
          example: 60
        sessionType:
          type: string
          enum: [initial, follow-up, emergency, final]
          default: follow-up
        notes:
          type: object
          properties:
            symptoms:
              type: string
            observations:
              type: string
            interventions:
              type: string
            homework:
              type: string
            progressNotes:
              type: string
        mood:
          type: object
          properties:
            before:
              type: integer
              minimum: 1
              maximum: 10
            after:
              type: integer
              minimum: 1
              maximum: 10
        goals:
          type: array
          items:
            type: string
        isCompleted:
          type: boolean
          default: false
    
    # Medical Record Schema
    MedicalRecord:
      type: object
      required:
        - patientId
        - recordType
        - title
        - addedBy
      properties:
        recordId:
          type: string
          readOnly: true
        patientId:
          type: string
        recordType:
          type: string
          enum: [diagnosis, medication, allergy, lab-result, history, other]
        title:
          type: string
        description:
          type: string
        diagnosis:
          type: object
          properties:
            code:
              type: string
              example: "F41.1"
            name:
              type: string
              example: "Generalized Anxiety Disorder"
            severity:
              type: string
              enum: [mild, moderate, severe]
        medications:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              dosage:
                type: string
              frequency:
                type: string
              startDate:
                type: string
                format: date
              endDate:
                type: string
                format: date
        allergies:
          type: array
          items:
            type: object
            properties:
              allergen:
                type: string
              reaction:
                type: string
              severity:
                type: string
                enum: [mild, moderate, severe, life-threatening]
        addedBy:
          type: string
        isActive:
          type: boolean
          default: true
    
    # Message Schema
    Message:
      type: object
      required:
        - senderId
        - senderType
        - receiverId
        - receiverType
        - content
      properties:
        messageId:
          type: string
          readOnly: true
        conversationId:
          type: string
        senderId:
          type: string
        senderType:
          type: string
          enum: [patient, therapist]
        receiverId:
          type: string
        receiverType:
          type: string
          enum: [patient, therapist]
        subject:
          type: string
        content:
          type: string
        messageType:
          type: string
          enum: [text, appointment-request, prescription, document, emergency]
          default: text
        priority:
          type: string
          enum: [low, normal, high, urgent]
          default: normal
        isRead:
          type: boolean
          default: false
    
    # Review Schema
    Review:
      type: object
      required:
        - therapistId
        - patientId
        - rating
      properties:
        reviewId:
          type: string
          readOnly: true
        therapistId:
          type: string
        patientId:
          type: string
        appointmentId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        categories:
          type: object
          properties:
            professionalism:
              type: integer
              minimum: 1
              maximum: 5
            communication:
              type: integer
              minimum: 1
              maximum: 5
            effectiveness:
              type: integer
              minimum: 1
              maximum: 5
            empathy:
              type: integer
              minimum: 1
              maximum: 5
        comment:
          type: string
        isAnonymous:
          type: boolean
          default: false
        isVerified:
          type: boolean
          default: false
    
    # Treatment Plan Schema
    TreatmentPlan:
      type: object
      required:
        - patientId
        - therapistId
        - title
        - startDate
      properties:
        planId:
          type: string
          readOnly: true
        patientId:
          type: string
        therapistId:
          type: string
        title:
          type: string
        description:
          type: string
        diagnosis:
          type: object
          properties:
            primary:
              type: string
            secondary:
              type: array
              items:
                type: string
        goals:
          type: array
          items:
            type: object
            properties:
              goalId:
                type: string
              description:
                type: string
              targetDate:
                type: string
                format: date
              status:
                type: string
                enum: [not-started, in-progress, achieved, abandoned]
              progress:
                type: integer
                minimum: 0
                maximum: 100
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, completed, on-hold, cancelled]
          default: active
    
    # Error Response
...
paths:
...
  /api/patients:
    post:
      tags:
        - Patients
      summary: Create a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Patient'
      responses:
        '201':
          description: Patient created successfully
        '400':
          description: Patient already exists
        '500':
          description: Server error
    
    get:
      tags:
        - Patients
      summary: Get all patients
      responses:
        '200':
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Patient'
        '500':
          description: Server error
...
  /api/therapists:
    post:
      tags:
        - Therapists
      summary: Create a new therapist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Therapist'
      responses:
        '201':
          description: Therapist created successfully
        '400':
          description: Therapist already exists
    
    get:
      tags:
        - Therapists
      summary: Get all therapists
      responses:
        '200':
          description: List of therapists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Therapist'
...
  /api/appointments:
    post:
      tags:
        - Appointments
      summary: Create a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
      responses:
        '201':
          description: Appointment created successfully
        '400':
          description: Time slot not available
    
    get:
      tags:
        - Appointments
      summary: Get appointments with filters
      parameters:
        - name: therapistId
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, completed, cancelled, rescheduled]
      responses:
        '200':
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
...
  /api/sessions:
    post:
      tags:
        - Sessions
      summary: Create a new therapy session record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Session'
      responses:
        '201':
          description: Session created successfully
    
    get:
      tags:
        - Sessions
      summary: Get sessions with filters
      parameters:
        - name: therapistId
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: string
        - name: appointmentId
          in: query
          schema:
            type: string
        - name: sessionType
          in: query
          schema:
            type: string
            enum: [initial, follow-up, emergency, final]
        - name: isCompleted
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
...
  /api/medical-records:
    post:
      tags:
        - Medical Records
      summary: Create a new medical record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicalRecord'
      responses:
        '201':
          description: Medical record created successfully
...
  /api/messages:
    post:
      tags:
        - Messages
      summary: Send a new message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '201':
          description: Message sent successfully
...
  /api/reviews:
    post:
      tags:
        - Reviews
      summary: Create a new review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
      responses:
        '201':
          description: Review created successfully
...
  /api/treatment-plans:
    post:
      tags:
        - Treatment Plans
      summary: Create a new treatment plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreatmentPlan'
      responses:
        '201':
          description: Treatment plan created successfully
...